Migrates send SSH key to SSH connect
